requirements.txt
email_validator
Flask
Flask-WTF
WTForms
cryptography
python-dotenv
PyMySQL
bcrypt
Werkzeug
Flask-Login
flask-limiter==3.7.0
itsdangerous
psycopg2-binary
pycryptodome
flask-bcrypt
pyjnius
kivy-garden
runnable
kivy
requests
pywebpush
sshtunnel
pymysql
pyarmor

gunicorn
